Update default configuration and related test

+++ src/bloom/command_handler.rs
@@ -111,7 +111,7 @@ pub fn bloom_filter_add_value(
             // Instantiate empty bloom filter.
             let fp_rate = configs::BLOOM_FP_RATE_DEFAULT;
             let tightening_ratio = configs::TIGHTENING_RATIO;
-            let capacity = configs::BLOOM_CAPACITY.load(Ordering::Relaxed) as u32;
+            let capacity = configs::BLOOM_CAPACITY.load(Ordering::Relaxed);
             let expansion = configs::BLOOM_EXPANSION.load(Ordering::Relaxed) as u32;
             let use_random_seed = configs::BLOOM_USE_RANDOM_SEED.load(Ordering::Relaxed);
             let mut bloom = match BloomFilterType::new_reserved(
@@ -233,7 +233,7 @@ pub fn bloom_filter_reserve(ctx: &Context, input_args: &[ValkeyString]) -> Valke
     };
     curr_cmd_idx += 1;
     // Parse the capacity
-    let capacity = match input_args[curr_cmd_idx].to_string_lossy().parse::<u32>() {
+    let capacity = match input_args[curr_cmd_idx].to_string_lossy().parse::<i64>() {
         Ok(num) if (BLOOM_CAPACITY_MIN..=BLOOM_CAPACITY_MAX).contains(&num) => num,
         Ok(0) => {
             return Err(ValkeyError::Str(utils::CAPACITY_LARGER_THAN_0));
@@ -317,7 +317,7 @@ pub fn bloom_filter_insert(ctx: &Context, input_args: &[ValkeyString]) -> Valkey
     idx += 1;
     let mut fp_rate = configs::BLOOM_FP_RATE_DEFAULT;
     let mut tightening_ratio = configs::TIGHTENING_RATIO;
-    let mut capacity = configs::BLOOM_CAPACITY.load(Ordering::Relaxed) as u32;
+    let mut capacity = configs::BLOOM_CAPACITY.load(Ordering::Relaxed);
     let mut expansion = configs::BLOOM_EXPANSION.load(Ordering::Relaxed) as u32;
     let mut nocreate = false;
     while idx < argc {
@@ -357,7 +357,7 @@ pub fn bloom_filter_insert(ctx: &Context, input_args: &[ValkeyString]) -> Valkey
                     return Err(ValkeyError::WrongArity);
                 }
                 idx += 1;
-                capacity = match input_args[idx].to_string_lossy().parse::<u32>() {
+                capacity = match input_args[idx].to_string_lossy().parse::<i64>() {
                     Ok(num) if (BLOOM_CAPACITY_MIN..=BLOOM_CAPACITY_MAX).contains(&num) => num,
                     Ok(0) => {
                         return Err(ValkeyError::Str(utils::CAPACITY_LARGER_THAN_0));
